Deploy/requirements.txt
Flask
flask_cors
tensorflow
numpy
Pillow